requirements.txt
python-telegram-bot==22.4
moviepy==2.2.1
python-dotenv==1.1.1